Sarkadi Péter's second commit. Makefile added.
