+++ switch.h
@@ -1,0 +1,5 @@
+#include <iostream>
+#include <string>
+using namespace std;
+
+string mondat(string proba);
